Object Pooling
+ Added Object Pooling
+ Game difficulty increases after time
+ various quality of life updates

+++ Assets/Scripts/Platform.cs
@@ -10,9 +10,9 @@ public class Platform : MonoBehaviour
     List<Transform> objectSpawnPositions;
 #pragma warning restore 649
 
-    int platformSpeed;
+    float platformSpeed;
 
-    public int PlatformSpeed { get { return platformSpeed; } set { platformSpeed = value; } }
+    public float PlatformSpeed { get { return platformSpeed; } set { platformSpeed = value; } }
 
     float obstaclePosition;
